+++ admin/class-decalog-admin.php
@@ -451,6 +451,7 @@ class Decalog_Admin {
 		if ( ! empty( $_POST ) ) {
 			if ( array_key_exists( '_wpnonce', $_POST ) && wp_verify_nonce( $_POST['_wpnonce'], 'decalog-plugin-options' ) ) {
 				Option::network_set( 'auto_update', array_key_exists( 'decalog_plugin_options_autoupdate', $_POST ) );
+				Option::network_set( 'use_cdn', array_key_exists( 'decalog_plugin_options_usecdn', $_POST ) );
 				Option::network_set( 'display_nag', array_key_exists( 'decalog_plugin_options_nag', $_POST ) );
 				Option::network_set( 'logger_autostart', array_key_exists( 'decalog_loggers_options_autostart', $_POST ) );
 				Option::network_set( 'pseudonymization', array_key_exists( 'decalog_loggers_options_pseudonymization', $_POST ) );
@@ -725,6 +726,22 @@ class Decalog_Admin {
 	 */
 	public function plugin_options_section_callback() {
 		$form = new Form();
+		add_settings_field(
+			'decalog_plugin_options_usecdn',
+			__( 'Resources', 'decalog' ),
+			[ $form, 'echo_field_checkbox' ],
+			'decalog_plugin_options_section',
+			'decalog_plugin_options_section',
+			[
+				'text'        => esc_html__( 'Use public CDN', 'decalog' ),
+				'id'          => 'decalog_plugin_options_usecdn',
+				'checked'     => Option::network_get( 'use_cdn' ),
+				'description' => esc_html__( 'Use CDN (jsDelivr) to serve DecaLog scripts and stylesheets.', 'decalog' ),
+				'full_width'  => true,
+				'enabled'     => true,
+			]
+		);
+		register_setting( 'decalog_plugin_options_section', 'decalog_plugin_options_usecdn' );
 		add_settings_field(
 			'decalog_plugin_options_autoupdate',
 			__( 'Plugin updates', 'decalog' ),