+++ admin/partials/decalog-admin-settings-about.php
@@ -11,6 +11,9 @@
 
 use Decalog\System\Environment;
 
+wp_enqueue_style( DECALOG_ASSETS_ID );
+wp_enqueue_script( DECALOG_ASSETS_ID );
+
 $warning = '';
 if ( Environment::is_plugin_in_dev_mode() ) {
 	$icon     = '<img style="width:16px;vertical-align:text-bottom;" src="' . \Feather\Icons::get_base64( 'alert-triangle', 'none', '#FF8C00' ) . '" />&nbsp;';
@@ -20,7 +23,7 @@ if ( Environment::is_plugin_in_rc_mode() ) {
 	$icon     = '<img style="width:16px;vertical-align:text-bottom;" src="' . \Feather\Icons::get_base64( 'alert-triangle', 'none', '#FF8C00' ) . '" />&nbsp;';
 	$warning .= '<p>' . $icon . sprintf( esc_html__( 'This version of %s is a release candidate. Although ready for production, this version is not officially supported in production environments.', 'decalog' ), DECALOG_PRODUCT_NAME ) . '</p>';
 }
-
+$icon       = '<img class="decalog-about-logo" style="opacity:0;" src="' . Decalog\Plugin\Core::get_base64_logo() . '" />';
 $intro      = sprintf( esc_html__( '%1$s is a free and open source plugin for WordPress. It integrates other free and open source works (as-is or modified) like: %2$s.', 'decalog' ), '<em>' . DECALOG_PRODUCT_NAME . '</em>', do_shortcode( '[decalog-libraries]' ) );
 $trademarks = esc_html__( 'All brands, icons and graphic illustrations are registered trademarks of their respective owners.', 'decalog' );
 $brands     = array( 'Automattic', 'Fluentd Project', 'Google', 'Pushover', 'Slack' );
@@ -28,6 +31,7 @@ $official   = sprintf( esc_html__( 'This plugin is not an official software from
 
 ?>
 <h2><?php echo esc_html( DECALOG_PRODUCT_NAME . ' ' . DECALOG_VERSION ); ?></h2>
+<?php echo $icon; ?>
 <?php echo $warning; ?>
 <p><?php echo $intro; ?></p>
 <h4><?php esc_html_e( 'Disclaimer', 'decalog' ); ?></h4>